// AIRMAM5/Scripts/Models/Interface/Subject/SubjectListModel.ts
/**主題列表 */
export interface SubjectListModel {
    /**主題編號 */
    fsSUBJECT_ID: string;
    /**標題 */
    fsSUBJECT_TITLE: string;
    /**影數量 */
    VideoCount: number;
    /**音數量 */
    AudioCount: number;
    /**圖數量 */
    PhotoCount: number;
    /**文數量 */
    DocCount: number;
}